fix(imgproc): address remap review — typed maps and module layout
Move remap_f32_cuda from interpolation/cuda.rs into interpolation/remap.rs
so the CUDA path sits beside the CPU path, and delete the now-empty module
along with its `pub(crate) mod cuda;` declaration. The device tests move
with it as a `cuda_tests` sibling module.

The adapter now takes `map_x: &Image<f32, 1>` / `map_y: &Image<f32, 1>`
instead of raw `&CudaSlice<f32>`; the `as_cudaslice()` extraction moved from
the call site into the adapter, leaving `remap` with only the residency
guard.

In calibration/distortion.rs, restore the two comments dropped when
generate_correction_map_polynomial was converted to return images, import
ImageError rather than spelling out kornia_image::ImageError, and fix the
doc comment that still described the old TensorError return.

Verified on a GTX 1650: the seven ignored CUDA remap tests pass, including
the bit-exact device-vs-host parity checks.
