+++ crates/kornia-imgproc/src/interpolation/mod.rs
@@ -1,7 +1,5 @@
 mod bicubic;
 mod bilinear;
-#[cfg(feature = "cuda")]
-pub(crate) mod cuda;
 pub(crate) mod lanczos;
 
 /// Utility functions to generate meshgrid and remap images
